requirements.txt
viam-sdk==0.41.0
python-dotenv==1.0.0